requirements.txt
numpy==1.24.0
opencv_python==4.6.0.66
Pillow==9.3.0
PyQt5==5.15.7
selenium==4.7.2
